+++ service/lesson.service.ts
@@ -1,5 +1,7 @@
+import { NotFoundError } from "@/errors/notfound";
+import { requireRole } from "@/lib/authorization";
 import { LessonRepository } from "@/repository/lesson.repository";
-import { LessonRequestDTO } from "@/types";
+import { LessonRequestDTO, TokenPayload } from "@/types";
 import { LessonUpdateRequestDTO } from "@/types";
 
 export class LessonService{
@@ -7,7 +9,12 @@ export class LessonService{
     
     constructor(){}
 
-    async createLesson(data: LessonRequestDTO){
+    async createLesson(data: LessonRequestDTO, user: TokenPayload){
+        requireRole(user, [
+            "ADMIN",
+            "INSTRUCTOR"
+        ])
+
         if(!data.title || !data.order){
             throw new Error("Dados obrigatórios não informados")
         }
@@ -23,13 +30,35 @@ export class LessonService{
         return lessons;
     }
 
-    async deleteLesson(id: string){
+    async deleteLesson(id: string, user: TokenPayload){
+        requireRole(user, [
+            "ADMIN",
+            "INSTRUCTOR"
+        ]);
+
+        const existingLesson = await this.lessonRepository.finById(id);
+
+        if(!existingLesson) {
+            throw new NotFoundError("Aula não encontrada")
+        }
+
         const lesson = await this.lessonRepository.delete(id);
 
         return lesson;
     }
 
-    async updateLesson(id: string, lesson: LessonUpdateRequestDTO) {
+    async updateLesson(id: string, lesson: LessonUpdateRequestDTO, user: TokenPayload) {
+        requireRole(user, [
+            "ADMIN",
+            "INSTRUCTOR"
+        ]);
+        
+        const existingLesson = this.lessonRepository.finById(id);
+
+        if(!existingLesson) {
+            throw new NotFoundError("Lição não encontrada");
+        }
+
         const updatedLesson = await this.lessonRepository.update(id, lesson);
 
         return updatedLesson;
